Switch to APIv3; search for tags on client
APIv2 isn't supported and the search is broken. GitHub suggested moving
to APIv3. Conveniently APIv3 doesn't have a search feature :(  So this
pulls down all the comments and searches them manually.

GitHub limits API calls to 5000 per hour per client. This could be an
issue with the number of issues open. In theory if I added oauth, they
could lift that limit.